Refactor: Simplify bus fare calculation logic and update `float` to `double` for fare and attendance variables.

+++ cse2124_w26/lab_3/exam_eligibility.c
@@ -11,11 +11,11 @@
 #include <stdio.h>
 
 int main() {
-    float attendance;
+    double attendance;
     int feesPaid;
 
     printf("Enter attendance percentage: ");
-    scanf("%f", &attendance);
+    scanf("%lf", &attendance);
 
     printf("Have you paid the fees? (1 for Yes, 0 for No): ");
     scanf("%d", &feesPaid);
